Refinement of the Perform Escape example.
Documentation comment for the class to identify which part of the book it is for.

Added titles to the two different modal view controllers so the user will have context for whether the “Perform Escape” action is available or not.

Changed the background colour for the “with Perform Escape” modal to `systemGreen`, to further distinguish it from the “without” modal.

Also ran the source against SwiftLint:
* removed trailing whitespace
* broke up lines > 120 characters long
* added bundle parameter to `NSLocalizedString` calls
* made `@IBAction` functions `private`