Add SwiftLint to build process.
SwiftLint, if available in the command path (Homebrew installed it as `/usr/local/bin/swiftlint` on my system), will be run as part of the Build process, as a step immediately prior to the compile step.

Added some documentation to the `README.md` file to explain how to install and configure SwiftLint.